+++ src/lib/enr/v4.ts
@@ -1,7 +1,8 @@
-import crypto from "crypto";
-
+import * as secp from "@noble/secp256k1";
 import { keccak256 } from "js-sha3";
-import * as secp256k1 from "secp256k1";
+
+import { randomBytes } from "../crypto";
+import { bytesToHex } from "../utils";
 
 import { createNodeId } from "./create";
 import { NodeId } from "./types";
@@ -10,17 +11,26 @@ export function hash(input: Uint8Array): Uint8Array {
   return new Uint8Array(keccak256.arrayBuffer(input));
 }
 
-export async function createPrivateKey(): Promise<Uint8Array> {
+export function createPrivateKey(): Uint8Array {
   return randomBytes(32);
 }
 
 export function publicKey(privKey: Uint8Array): Uint8Array {
-  return secp256k1.publicKeyCreate(privKey);
+  return secp.getPublicKey(privKey, true);
 }
 
-export function sign(privKey: Uint8Array, msg: Uint8Array): Uint8Array {
-  const { signature } = secp256k1.ecdsaSign(hash(msg), privKey);
-  return signature;
+export function compressPublicKey(publicKey: Uint8Array): Uint8Array {
+  const point = secp.Point.fromHex(bytesToHex(publicKey));
+  return point.toRawBytes(true);
+}
+
+export async function sign(
+  privKey: Uint8Array,
+  msg: Uint8Array
+): Promise<Uint8Array> {
+  return secp.sign(hash(msg), privKey, {
+    der: false,
+  });
 }
 
 export function verify(
@@ -28,12 +38,17 @@ export function verify(
   msg: Uint8Array,
   sig: Uint8Array
 ): boolean {
-  // Remove the recovery id if present (byte #65)
-  return secp256k1.ecdsaVerify(sig.slice(0, 64), hash(msg), pubKey);
+  try {
+    const _sig = secp.Signature.fromCompact(sig.slice(0, 64));
+    return secp.verify(_sig, hash(msg), pubKey);
+  } catch {
+    return false;
+  }
 }
 
 export function nodeId(pubKey: Uint8Array): NodeId {
-  const uncompressedPubkey = secp256k1.publicKeyConvert(pubKey, false);
+  const publicKey = secp.Point.fromHex(pubKey);
+  const uncompressedPubkey = publicKey.toRawBytes(false);
 
   return createNodeId(hash(uncompressedPubkey.slice(1)));
 }
@@ -45,34 +60,24 @@ export class ENRKeyPair {
     public readonly publicKey: Uint8Array
   ) {}
 
-  public static async create(privateKey?: Uint8Array): Promise<ENRKeyPair> {
+  public static create(privateKey?: Uint8Array): ENRKeyPair {
     if (privateKey) {
-      if (!secp256k1.privateKeyVerify(privateKey)) {
+      if (!secp.utils.isValidPrivateKey(privateKey)) {
         throw new Error("Invalid private key");
       }
     }
-    const _privateKey = privateKey || (await createPrivateKey());
+    const _privateKey = privateKey || createPrivateKey();
     const _publicKey = publicKey(_privateKey);
     const _nodeId = nodeId(_publicKey);
 
     return new ENRKeyPair(_nodeId, _privateKey, _publicKey);
   }
 
-  public sign(msg: Uint8Array): Uint8Array {
+  public async sign(msg: Uint8Array): Promise<Uint8Array> {
     return sign(this.privateKey, msg);
   }
 
   public verify(msg: Uint8Array, sig: Uint8Array): boolean {
     return verify(this.publicKey, msg, sig);
-  }
-}
-
-function randomBytes(length: number): Uint8Array {
-  if (typeof window !== "undefined" && window && window.crypto) {
-    const array = new Uint8Array(length);
-    window.crypto.getRandomValues(array);
-    return array;
-  } else {
-    return crypto.randomBytes(length);
   }
 }